Add Parcel and initial html file. Move contents of Game class into Board
