+++ package.json
@@ -4,7 +4,9 @@
   "description": "Game of Life implementation in javascript and html5 canvas",
   "main": "index.js",
   "scripts": {
-    "test": "jasmine"
+    "test": "jasmine",
+    "dev": "parcel index.html",
+    "build": "parcel build index.html"
   },
   "author": "Colin Hulme",
   "license": "GPL-2.0",
@@ -14,7 +16,8 @@
     "eslint-plugin-import": "^2.16.0",
     "eslint-plugin-jasmine": "^2.10.1",
     "jasmine": "^3.4.0",
-    "jasmine-node": "^1.16.2"
+    "jasmine-node": "^1.16.2",
+    "parcel-bundler": "^1.12.3"
   },
   "dependencies": {}
 }
