+++ spec/support/board_spec.js
@@ -1,14 +1,14 @@
-const Game = require('../../src/game');
+const Board = require('../../src/board');
 
 
-describe('Game', () => {
-  let game;
+describe('Board', () => {
+  let board;
   let cell;
   let liveCell;
   let subjectCell;
 
   beforeEach(() => {
-    game = new Game(5);
+    board = new Board(5);
     cell = jasmine.createSpyObj('cell', ['isAlive']);
     cell.isAlive.and.callFake(() => false);
 
@@ -21,23 +21,23 @@ describe('Game', () => {
 
   describe('#allDead', () => {
     it('returns true for a board full of dead cells', () => {
-      game.board = [[cell, cell], [cell, cell]];
+      board.board = [[cell, cell], [cell, cell]];
 
-      expect(game.allDead()).toBe(true);
+      expect(board.allDead()).toBe(true);
     });
 
     it('returns false for a board with 1 or more alive cells', () => {
-      game.board = [[cell, cell], [cell, liveCell]];
+      board.board = [[cell, cell], [cell, liveCell]];
 
-      expect(game.allDead()).toBe(false);
+      expect(board.allDead()).toBe(false);
     });
   });
 
   describe('#getCell', () => {
     it('returns the cell at given x and y coordinates', () => {
-      game.board = [[cell, cell], [cell, subjectCell]];
+      board.board = [[cell, cell], [cell, subjectCell]];
 
-      expect(game.getCell(1, 1)).toEqual(subjectCell);
+      expect(board.getCell(1, 1)).toEqual(subjectCell);
     });
   });
 
@@ -53,12 +53,12 @@ describe('Game', () => {
     const testCells = [cell1, cell2, cell3, cell4, cell5, cell6, cell7, cell8];
 
     it('identified all cells that are neighbours', () => {
-      game.board = [[cell1, cell2, cell3], [cell4, subjectCell, cell5], [cell6, cell7, cell8]];
+      board.board = [[cell1, cell2, cell3], [cell4, subjectCell, cell5], [cell6, cell7, cell8]];
       for (let i = 0; i < testCells.length; i += 1) {
-        expect(game.getNeighbours(1, 1)).toContain(testCells[i]);
+        expect(board.getNeighbours(1, 1)).toContain(testCells[i]);
       }
-      expect(game.getNeighbours(1, 1).length).toBe(8);
-      expect(game.getNeighbours(1, 1)).not.toContain(subjectCell);
+      expect(board.getNeighbours(1, 1).length).toBe(8);
+      expect(board.getNeighbours(1, 1)).not.toContain(subjectCell);
     });
 
     xit('can return correct results for non-central pieces', () => {
